Get weight tickets list

+++ apps/springboot-backend/src/main/kotlin/nl/eazysoftware/eazyrecyclingservice/domain/weightticket/WeightTicket.kt
@@ -29,5 +29,5 @@ enum class WeightTicketStatus {
 }
 
 data class WeightTicketId(
-  val number: Int
+  val number: Long
 )
